+++ Example/ReusableDemo/CollectionViewCells/MyColorSquareCell.swift
@@ -21,13 +21,13 @@ import Reusable
 final class MyColorSquareCell: UICollectionViewCell, Reusable {
   private lazy var colorView: UIView = {
     let colorView = UIView()
-    colorView.autoresizingMask = [.FlexibleWidth, .FlexibleHeight]
+    colorView.autoresizingMask = [.flexibleWidth, .flexibleHeight]
     colorView.frame = self.contentView.bounds.insetBy(dx: 10, dy: 10)
     self.contentView.addSubview(colorView)
     return colorView
   }()
 
-  func fill(color: UIColor) {
+  func fill(_ color: UIColor) {
     self.colorView.backgroundColor = color
   }
 }