+++ Example/ReusableDemo/CollectionViewCells/MyXIBIndexSquaceCell.swift
@@ -24,7 +24,7 @@ final class MyXIBIndexSquaceCell: UICollectionViewCell, NibReusable {
   @IBOutlet private weak var sectionLabel: UILabel!
   @IBOutlet private weak var rowLabel: UILabel!
 
-  func fill(indexPath: NSIndexPath) {
+  func fill(_ indexPath: IndexPath) {
     sectionLabel.text = String(indexPath.section)
     rowLabel.text = String(indexPath.row)
   }