+++ Example/ReusableDemo/TableViewCells/MyXIBTextCell.swift
@@ -24,7 +24,7 @@ final class MyXIBTextCell: UITableViewCell, NibReusable {
 
   @IBOutlet private weak var titleLabel: UILabel!
 
-  func fill(text: String) {
+  func fill(_ text: String) {
     titleLabel.text = text
   }
 }