+++ Sources/View/UITableView+Reusable.swift
@@ -16,10 +16,10 @@ public extension UITableView {
 
    - parameter cellType: the `UITableViewCell` (`NibReusable`-conforming) subclass to register
 
-   - seealso: `registerNib(_:,forCellReuseIdentifier:)`
+   - seealso: `register(_:,forCellReuseIdentifier:)`
    */
-  final func registerReusableCell<T: UITableViewCell where T: NibReusable>(cellType: T.Type) {
-    self.registerNib(cellType.nib, forCellReuseIdentifier: cellType.reuseIdentifier)
+  final func register<T: UITableViewCell>(cellType: T.Type) where T: NibReusable {
+    self.register(cellType.nib, forCellReuseIdentifier: cellType.reuseIdentifier)
   }
 
   /**
@@ -27,10 +27,10 @@ public extension UITableView {
 
    - parameter cellType: the `UITableViewCell` (`Reusable`-conforming) subclass to register
 
-   - seealso: `registerClass(_:,forCellReuseIdentifier:)`
+   - seealso: `register(_:,forCellReuseIdentifier:)`
    */
-  final func registerReusableCell<T: UITableViewCell where T: Reusable>(cellType: T.Type) {
-    self.registerClass(cellType.self, forCellReuseIdentifier: cellType.reuseIdentifier)
+  final func register<T: UITableViewCell>(cellType: T.Type) where T: Reusable {
+    self.register(cellType.self, forCellReuseIdentifier: cellType.reuseIdentifier)
   }
 
   /**
@@ -43,10 +43,10 @@ public extension UITableView {
 
    - note: The `cellType` parameter can generally be omitted and infered by the return type,
    except when your type is in a variable and cannot be determined at compile time.
-   - seealso: `dequeueReusableCellWithIdentifier(_:,forIndexPath:)`
+   - seealso: `dequeueReusableCell(withIdentifier:,for:)`
    */
-  final func dequeueReusableCell<T: UITableViewCell where T: Reusable>(indexPath indexPath: NSIndexPath, cellType: T.Type = T.self) -> T {
-    guard let cell = self.dequeueReusableCellWithIdentifier(cellType.reuseIdentifier, forIndexPath: indexPath) as? T else {
+  final func dequeueReusableCell<T: UITableViewCell>(for indexPath: IndexPath, cellType: T.Type = T.self) -> T where T: Reusable {
+    guard let cell = self.dequeueReusableCell(withIdentifier: cellType.reuseIdentifier, for: indexPath) as? T else {
       fatalError(
         "Failed to dequeue a cell with identifier \(cellType.reuseIdentifier) matching type \(cellType.self). "
           + "Check that the reuseIdentifier is set properly in your XIB/Storyboard "
@@ -59,23 +59,23 @@ public extension UITableView {
   /**
    Register a NIB-Based `UITableViewHeaderFooterView` subclass (conforming to `NibReusable`)
 
-   - parameter viewType: the `UITableViewHeaderFooterView` (`NibReusable`-conforming) subclass to register
+   - parameter headerFooterViewType: the `UITableViewHeaderFooterView` (`NibReusable`-conforming) subclass to register
 
-   - seealso: `registerNib(_:,forHeaderFooterViewReuseIdentifier:)`
+   - seealso: `register(_:,forHeaderFooterViewReuseIdentifier:)`
    */
-  final func registerReusableHeaderFooterView<T: UITableViewHeaderFooterView where T: NibReusable>(viewType: T.Type) {
-    self.registerNib(viewType.nib, forHeaderFooterViewReuseIdentifier: viewType.reuseIdentifier)
+  final func register<T: UITableViewHeaderFooterView>(headerFooterViewType: T.Type) where T: NibReusable {
+    self.register(headerFooterViewType.nib, forHeaderFooterViewReuseIdentifier: headerFooterViewType.reuseIdentifier)
   }
 
   /**
    Register a Class-Based `UITableViewHeaderFooterView` subclass (conforming to `Reusable`)
 
-   - parameter viewType: the `UITableViewHeaderFooterView` (`Reusable`-confirming) subclass to register
+   - parameter headerFooterViewType: the `UITableViewHeaderFooterView` (`Reusable`-confirming) subclass to register
 
-   - seealso: `registerClass(_:,forHeaderFooterViewReuseIdentifier:)`
+   - seealso: `register(_:,forHeaderFooterViewReuseIdentifier:)`
    */
-  final func registerReusableHeaderFooterView<T: UITableViewHeaderFooterView where T: Reusable>(viewType: T.Type) {
-    self.registerClass(viewType.self, forHeaderFooterViewReuseIdentifier: viewType.reuseIdentifier)
+  final func register<T: UITableViewHeaderFooterView>(headerFooterViewType: T.Type) where T: Reusable {
+    self.register(headerFooterViewType.self, forHeaderFooterViewReuseIdentifier: headerFooterViewType.reuseIdentifier)
   }
 
   /**
@@ -87,10 +87,10 @@ public extension UITableView {
 
    - note: The `viewType` parameter can generally be omitted and infered by the return type,
    except when your type is in a variable and cannot be determined at compile time.
-   - seealso: `dequeueReusableHeaderFooterViewWithIdentifier(_:)`
+   - seealso: `dequeueReusableHeaderFooterView(withIdentifier:)`
    */
-  final func dequeueReusableHeaderFooterView<T: UITableViewHeaderFooterView where T: Reusable>(viewType: T.Type = T.self) -> T? {
-    guard let view = self.dequeueReusableHeaderFooterViewWithIdentifier(viewType.reuseIdentifier) as? T? else {
+  final func dequeueReusableHeaderFooterView<T: UITableViewHeaderFooterView>(_ viewType: T.Type = T.self) -> T? where T: Reusable {
+    guard let view = self.dequeueReusableHeaderFooterView(withIdentifier: viewType.reuseIdentifier) as? T? else {
       fatalError(
         "Failed to dequeue a header/footer with identifier \(viewType.reuseIdentifier) matching type \(viewType.self). "
           + "Check that the reuseIdentifier is set properly in your XIB/Storyboard "